Arabic files and translation
We might have trouble to the right to left format in Arabic, hebrew etc. I did a light googling https://stackoverflow.com/questions/2258028/rtl-in-markdown
Here is one seems-to-simple suggestion:

"""Actually as my friend Aevyz reminded me, Markdown parses HTML in it.

You won't need to change your parser. The quickest path to solve that I could think of is this:

<div dir="rtl">

سلام دنیا

مرحبا العالم

שלום בעולם

ہیلو دنیا
</div>
So you need to add literally two lines to turn a whole document RTL, and by the way it will be better compatible than an own script. So i think this is an answer to the question.""""